Create and diagonalize mid-circuit measurements in conditionals (#7037)
**Context:**

- Currently, we can't correctly execute mid-circuit measurements inside
`Conditional`s, so we need to get rid of those before we get to
execution.
- The wrapper created in `qml.cond` doesn't return anything (so its not
suitable for conditionally applying functions that return a
`MeasurementValue`), and would need different handling for measurement
functions than for an `Operator` or `qfunc`. We've decided to continue
limiting it `Operator` and `qfunc`, and make a separate function just
for measurements. It will be restricted in its inputs and outputs to
make sure that what it returns is up-to-date with our current ability to
process conditional MCMs, and it will be part of the `ftqc` module.

**Description of the Change:**

- We add `cond_meas`, with a call signature `condition, true_fn,
false_fn,` to match `qml.cond`. Limitation on inputs will be informed by
our limitations on what we can process for execution, and will currently
be:

- both `true_fn` and `false_fn` are required, and should create a
`MidMeasureMP` (likely a parametrized one from the `ftqc` module) and
return a `MeasurementValue`
- the measurements on both branches must be applied to the same wire,
and they must have the same `reset` and `postselect` attributes.
Effectively this means only the measurement basis is conditional.

- We update the `diagonalize_mcms` transform to assume any `Conditional`
containing a mid-circuit measurement is accompanied by its corresponding
`false_fn` `Conditional` on the tape, and to diagonalize pairs of
`Conditional`s together. The transform "factors out" the measurement as
a Z-basis measurement, so it's no longer conditional, and applies
conditional diagonalizing gates.

**Benefits:**

We can add a measurement to a circuit whose measurement basis is
conditional on a previous measurement outcome, and retrieve the
measurement value.

**Possible Drawbacks:**
It's restrictive, and it's limited to devices where we want to
diagonalize the MCMs - but for now, it does what we need it to do.

[sc-84499]

---------

Co-authored-by: Lee James O'Riordan <mlxd@users.noreply.github.com>
Co-authored-by: Joey Carter <joseph.carter@xanadu.ai>