Remove redundant overload for MCM mod (#7085)
### Before submitting

Please complete the following checklist when submitting a PR:

- [x] All new features must include a unit test.
If you've fixed a bug or added code that should be tested, add a test to
the
      test directory!

- [x] All new functions and code must be clearly commented and
documented.
If you do make documentation changes, make sure that the docs build and
      render correctly by running `make docs`.

- [x] Ensure that the test suite passes, by running `make test`.

- [x] Add a new entry to the `doc/releases/changelog-dev.md` file,
summarizing the
      change, and including a link back to the PR.

- [x] The PennyLane source code conforms to
      [PEP8 standards](https://www.python.org/dev/peps/pep-0008/).
We check all of our code against [Pylint](https://www.pylint.org/).
      To lint modified files, simply `pip install pylint`, and then
      run `pylint pennylane/path/to/file.py`.

When all the above are checked, delete everything above the dashed
line and fill in the pull request template.


------------------------------------------------------------------------------------------------------------

**Context:** PR https://github.com/PennyLaneAI/pennylane/pull/6981 added
support for MOD and XOR with MCMs. An incorrect offload name was
assigned to `mod`, which defined an autoray dispatch as `logical_mod`.
This naming is nonsensical, and causes issues with dispatch due to the
disparate naming from the numpy APIs `numpy.mod`, such as
`ImportError: autoray couldn't find function 'logical_mod' for backend
'numpy'.` when called using LightningQubit.

**Description of the Change:** This PR reassigns the MCM `__mod__`
dunder to call into `qml.math.mod`, which should implicitly work with
the outcome of measures and offloads to native frameworks, be they
built-in or numpy.

**Benefits:** Fixes error observed with executing dynamic MCM workloads
with LightningQubit, due to the naming and data type mismatch.

**Possible Drawbacks:**

**Related GitHub Issues:**