[Dynamic Decomposition] Fix single wire `QFT` decomposition (#7081)
**Context:**

Prior to this PR, we had the behaviour,
```python
import jax
qml.capture.enable()

@qml.capture.expand_plxpr_transforms
@qml.transforms.decompose
def circuit():
    # Should decompose to just H
    qml.QFT(wires=[0])

>>> jax.make_jaxpr(circuit)()
IndexError: index is out of bounds for axis 0 with size 0
```

**Description of the Change:**

Problem was that in the decomposition,
```python
...
       @qml.for_loop(n_wires)
        def outer_loop(i):
            qml.Hadamard(wires[i])

            @qml.for_loop(shift_len - i)
            def cphaseshift_loop(j):
                qml.ControlledPhaseShift(shifts[j], wires=[wires[i + j + 1], wires[i]])

            cphaseshift_loop()

        outer_loop()
...
```
The variable `shift_len` could be zero if the template is applied to one
wire. Therefore, the fix is just to check if we are dealing with one
wire or not.

**Benefits:** Code works for all cases.

**Possible Drawbacks:** None identified.

[sc-86417]

---------

Co-authored-by: Joseph Lee <40768758+josephleekl@users.noreply.github.com>