[Capture] extract register_custom_staging_rule (#7086)
**Context:**

When repurposing the logic introduced #6888 from working with just
`cond` to the`while_loop`, I found that I could in fact abstract the
logic. This PR extracts that logic.

**Description of the Change:**

Extracts a `qml.capture.register_custom_staging_rule` function from
being private for `cond`, to generic for more higher order primitives.

**Benefits:**

Less duplicate code.

**Possible Drawbacks:**

While this abstraction does work for both cond and while, it may break
in other situations, and probably undergo a lot of changes in the
future.

**Related GitHub Issues:**

This extract logic from #7084 to mimize the moving pieces in that PR.
[sc-86474]

---------

Co-authored-by: Yushao Chen (Jerry) <chenys13@outlook.com>
Co-authored-by: Pietropaolo Frisoni <pietropaolo.frisoni@xanadu.ai>