[Capture] Allow dynamic shapes to match other arguments (#7059)
**Context:**

If the size of a dynamic axis matches an argument already encountered,
jax doesn't add it to the start of the jaxpr's `invars`.

For example, with

```
def f(i, x):
    return 

def w(i):
    return f(i, jax.numpy.arange(i))
```

**Description of the Change:**

Patches the higher order primitives to handle the above case.  

**Benefits:**

We can handle accepting functions like the above.

**Possible Drawbacks:**

The current method of handling this for `for_loop` and `while_loop` is
ugly and patchy. I will try and get better abstractions once more
functionality and cases has been added.

**Related GitHub Issues:**

[sc-85871]

---------

Co-authored-by: Pietropaolo Frisoni <pietropaolo.frisoni@xanadu.ai>
Co-authored-by: Andrija Paurevic <46359773+andrijapau@users.noreply.github.com>