Remove outdated Observable check in qml.equal (#7107)
**Context:**

If the first operation in `qml.equal` is an adjoint of an Observable, it
skips the "do they match" check because of an outdated clause that was
meant to allow comparison of instances of the removed `Tensor` and
`Hamiltonian` classes. This means that the second operation is not
confirmed to also be an `Adjoint` before jumping to the function that
assumes this check has already been performed, and leads to bugs like
this:

```
>>> qml.equal(qml.adjoint(H(0)), qml.Y(0))
AttributeError: 'PauliY' object has no attribute 'base'
```

and

```
>>> qml.equal(qml.adjoint(X(0)), qml.exp(qml.X(0), 2))
True
```

This came up in a comparison of to operations in a decomposition when
updating `assert_valid` (#7099).

**Description of the Change:**

Removed the exception for if `op1` is an `Observable`.

**Benefits:**

Fixes this edge-case bug for comparing operators where `op1` is the
adjoint of an observable, and `op2` is not an `Adjoint`.

**Possible Drawbacks:**

If the tests and I are both missing some edge case where this is still
relevant, I guess we might be breaking something.


[sc-86861]