Update decomposition check in assert_valid (#7099)
**Context:**
We have a check in `assert_valid` that ensures that an operator
decomposition doesn't include the same type of operator again. This is a
relevant check for operators types where a given class will be either
supported or not by devices - in this case, a decomposition that returns
another instance of the same type of operation will lead to an infinite
recursion.

However, for some operators, this check doesn't make sense.
Specifically, an operator with another operation as its base may
decompose based on the base operator, leaving the overall operator type
intact (for example, `Controlled` operations with an unsupported base).

This has come up in PR #6921 for the `FromBloq` operator.

**Description of the Change:**

- We check that the operator _itself_ is not present in its own
decomposition in all cases.
- We keep the check that the decomposition doesn't include any operators
of the same type as the original operator, but make it optional via a
new argument in `assert_valid`.
- We add an error message with instructions for how to skip that part of
the decomposition check, and when it would be valid to do so.

**Benefits:**
More flexibility for asserting validity of the group of operators where
specific instances may decompose into other operators of the same type.

**Possible Drawbacks:**
- We're starting to add a lot of different kinds of exceptions to
`assert_valid`.
- The added argument is currently called `decomposition_to_same_class`,
which is long and not that helpful, but I'm coming up blank when I try
to think of something better


[sc-86860]