[Decomposition] Defines the interface for decomposition rules (#6950)
**Context:**

This PR defines the data structure used to define a decomposition rule
and the API for creating a decomposition rule and registering it with an
operator. Many features and concepts are adapted from the resource
estimation module.

**Description of the Change:**

Created three classes:

- The `DecompositionRule` stores a decomposition rule as a qfunc and a
resource estimate.
- The `Resources` is a data structure to represent the resource estimate
of a decompositon.
- The `CompressedResourceOp` is a lightweight representation of an
operator to be decomposed.

Created functions:
- The `register_resources` decorator can be used to wrap a qfunc in a
`DecompositionRule` object.
- A `resource_rep` to create a `CompressedResourceOp` from an operator
type and any additional information.

Added methods and attributes to `Operator`:
- Class property `Operator.resource_keys` is a set of strings
- Properties `resource_params` and `resource_rep` that returns a
`CompressedReourceOp`

**Related GitHub Issues:**
[sc-84328]

---------

Co-authored-by: Jay Soni <jbsoni@uwaterloo.ca>