Reserve ABC for abstract base classes in codebase (#7156)
### Before submitting

Please complete the following checklist when submitting a PR:

- [x] All new features must include a unit test.
If you've fixed a bug or added code that should be tested, add a test to
the
      test directory!

- [x] All new functions and code must be clearly commented and
documented.
If you do make documentation changes, make sure that the docs build and
      render correctly by running `make docs`.

- [x] Ensure that the test suite passes, by running `make test`.

- [x] Add a new entry to the `doc/releases/changelog-dev.md` file,
summarizing the
      change, and including a link back to the PR.

- [x] The PennyLane source code conforms to
      [PEP8 standards](https://www.python.org/dev/peps/pep-0008/).
We check all of our code against [Pylint](https://www.pylint.org/).
      To lint modified files, simply `pip install pylint`, and then
      run `pylint pennylane/path/to/file.py`.

When all the above are checked, delete everything above the dashed
line and fill in the pull request template.


------------------------------------------------------------------------------------------------------------

**Context:** Searching and filtering of implementations can be
simplified if a convention is maintained for naming specific entities.
Given the standard practice of `import abc; abc.ABC` being used for
Pythonic abstract base classes, it can be challenging to filter abstract
classes under search when the same `ABC` naming is used elsewhere. This
PR ensures the original ASCII lettering users of the `ABC` label no
longer do so.

**Description of the Change:** As above.

**Benefits:** This allows a simplified approach for analysing the number
of abstract base classes within PennyLane through search and filtering
of the `ABC` label.

**Possible Drawbacks:** None.

**Related GitHub Issues:**

---------

Co-authored-by: David Wierichs <david.wierichs@xanadu.ai>