Add caching of null.qubit zero result generation (#7155)
### Before submitting

Please complete the following checklist when submitting a PR:

- [x] All new features must include a unit test.
If you've fixed a bug or added code that should be tested, add a test to
the
      test directory!

- [x] All new functions and code must be clearly commented and
documented.
If you do make documentation changes, make sure that the docs build and
      render correctly by running `make docs`.

- [x] Ensure that the test suite passes, by running `make test`.

- [x] Add a new entry to the `doc/releases/changelog-dev.md` file,
summarizing the
      change, and including a link back to the PR.

- [x] The PennyLane source code conforms to
      [PEP8 standards](https://www.python.org/dev/peps/pep-0008/).
We check all of our code against [Pylint](https://www.pylint.org/).
      To lint modified files, simply `pip install pylint`, and then
      run `pylint pennylane/path/to/file.py`.

When all the above are checked, delete everything above the dashed
line and fill in the pull request template.


------------------------------------------------------------------------------------------------------------

**Context:** As `null.qubit` is designed to mock device execution, it's
return of zero measurement data does not require unique generated
values. This PR adds intermediate caching of the zero values to improve
runtime and memory usage, avoiding the need of extra work or memory
allocations to mimic output.

**Description of the Change:** As above.

**Benefits:** Improves the memory usage and runtime of null.qubit with
large numbers of executions and/or MCMs.

**Possible Drawbacks:** Potentially less real-world device like in
execution.

**Related GitHub Issues:**