Create a custom label method for `PrepSelPrep` (#7164)
**Context:**
`PrepSelPrep` prints the coefficients of the encoded LCU as parameters
in its label.
That's very verbose and impractical, see #7159.

**Description of the Change:**
Implement a custom `label` method for `PrepSelPrep` that defers the
coefficients to the standard matrix cache,
which is also used for all matrix-valued parameters.
Like requested in the linked issue.

**Benefits:**
More readable drawings.

**Possible Drawbacks:**
N/A

**Related GitHub Issues:**
Resolves #7159

[sc-87444]

---------

Co-authored-by: Isaac De Vlugt <34751083+isaacdevlugt@users.noreply.github.com>