Adding Read the docs check (#7157)
------------------------------------------------------------------------------------------------------------

**Context:**
Pennylane previously implemented GitHub merge queues but faced issues
with the readthedocs step failing. Merge queues increase velocity by
automating pull request merges to a busy branch and ensuring the branch
is never broken by incompatible changes.

The merge queue provides the same benefits as the Require branches to be
up to date before merging branch protection, but does not require a pull
request author to update their pull request branch and wait for status
checks to finish before trying to merge.

With merge queues enabled the external Readthedocs does not run on pull
requests so the pipelines would not check if the documents are
buildable.

**Description of the Change:**
This change adds a pipeline that runs the same steps as the readthedocs
pipeline.

**Benefits:**
Running this check ensures that the readthedocs step will not fail at
the time of the PR's merge.

**Possible Drawbacks:**
There could be rare times when this new pipeline causes false positives
that lead to the readthedocs step failing.

**Related GitHub Issues:**

---------

Co-authored-by: Rashid N H M <95639609+rashidnhm@users.noreply.github.com>