Fix misc issues involving sparse related dispatches (#7140)
**Context:**
In a recent
[PR](https://github.com/PennyLaneAI/pennylane/pull/6963/files#diff-041bc8dd979e4e4c25bcc0670a480ec205320898b7e5c0b06fcba12e840ca040)
we introduced the dispatches for `allclose` to compare between
scipy.sparse obj and numpy. However, under some circumstances we also
need the compare between scipy.sparse and scalar, which has been missed
and caused unexpected
[failures](https://github.com/PennyLaneAI/pennylane-lightning/actions/runs/14035024064/job/39291431516?pr=1094#step:16:79),
therefore we introduce several more here to fix this issue.

Also, other issue was found during the debugging. The major one is the
incomplete multidispatch of `dot` when containing sparse matrices. It's
a SciPy issue mentioned
[here](https://github.com/scipy/scipy/issues/18938) that were able to be
avoided by using `@` instead with carefully treating the scalar cases.

**Description of the Change:**

**Benefits:**

**Possible Drawbacks:**

**Related GitHub Issues:**
https://github.com/scipy/scipy/issues/18938
[sc-87592]