Reuse lines for classical wires if possible (#7163)
**Context:**
As described in #7118 , a lot of white space is produced by the circuit
drawer if there are many MCMs, even if their impact is localized, and
the same lines below the quantum wires could be reused for different
classical wires.

**Description of the Change:**
This PR changes the classical wire allocation in an additional
processing step, reusing lines for classical wires if possible.

**Benefits:**
More concise circuit drawings.

**Possible Drawbacks:**
A little more processing happening. Worst case should be something like
$\mathcal{O}(k^2)$ scaling, where $k$ is the number of classical wires.
However, most operations are $\mathcal{O}(k)$ and the only
$\mathcal{O}(k^2)$ operation uses `np.where` as an inner loop, which
should come with very small constants.

**Related GitHub Issues:**
Resolves #7118 

[sc-86882]

---------

Co-authored-by: Mudit Pandey <mudit.pandey@xanadu.ai>
Co-authored-by: Christina Lee <christina@xanadu.ai>