Add `max_length` to `draw_mpl` (#7128)
**Context:**

Deep circuits can be rather hard to use with `draw_mpl`, as the image
just becomes way too large. So we are adding in a way of splitting

**Description of the Change:**

Adds a `max_length` kwarg to `draw_mpl` and `tape_mpl`, splitting them
into multiple figures if they go over that length.

**Benefits:**

Easier to visualize huge circuits.

**Possible Drawbacks:**

**Related GitHub Issues:**

[sc-84979]

---------

Co-authored-by: Yushao Chen (Jerry) <chenys13@outlook.com>
Co-authored-by: AntonNI8 <anton.naimibrahim@xanadu.ai>