HT gradient - additional modes (#7046)
**Context:** Hadamard gradients method variants

**Description of the Change:** added support for different modes in
hadamard gradient and implemented direct HT, reversed HT and
reversed_direct HT, whose details could be found at
https://arxiv.org/pdf/2408.05406

**Benefits:** more variants of HT + measurement grouping generates
better efficiency for gradient estimation

**Possible Drawbacks:** for reversed methods, that is reversed HT and
reversed_direct HT, the current version does not support circuits with
multiple measurements. One will need to modify the processing_fn in how
it aggregates the results so support this. For non-reversed methods, we
aggregate for each measurement and the code work because the shape is
N_G*M where N_G is the number of terms in the generator and M is the
number of measurements. However, in reversed method, the shape is [N_1,
..., N_M], where for the i-th measurement, there are N_i gradient tapes.
Different measurement has different number of tapes, and all the tape
has only one measurement, so one does not need to aggregate across
dimensions, just group neighboring ones would do.

**Related GitHub Issues:**

---------

Co-authored-by: Isaac De Vlugt <34751083+isaacdevlugt@users.noreply.github.com>
Co-authored-by: Isaac De Vlugt <isaacdevlugt@gmail.com>
Co-authored-by: albi3ro <chrissie.c.l@gmail.com>
Co-authored-by: Christina Lee <christina@xanadu.ai>
Co-authored-by: Yushao Chen (Jerry) <chenys13@outlook.com>
Co-authored-by: Alex Preciado <alex.preciado@xanadu.ai>