[Capture] User transforms are expanded before execution (#7199)
[sc-83799]

**Context:**
This story adds support for expanding user transforms before sending
jaxpr to devices when executing a QNode.

**Description of the Change:**
* Update `qnode_prim.impl` to expand any present user transforms in the
`qfunc_jaxpr` before applying device preprocessing transforms, using
`qml.capture.expand_plxpr_transforms`.
* Update `TransformDispatcher` to treat QNodes differently from other
callables when program capture is enabled. When transforming a `QNode`,
the transform will also be added to the `QNode`'s tranform program. This
does not change the captured program, but it gives us the ability to use
the transform program to apply qnode transforms to the `qfunc_jaxpr`.
* Update `transform_prim.impl` to just return the evaluation of the
inner jaxpr. This was done so that calling a transformed qnode (which is
actually a wrapper function around the qnode) will not raise a
`NotImplementedError`.

**Benefits:**
Users no longer need to manually decorate their qnode with
`@qml.capture.expand_plxpr_transforms`.

**Possible Drawbacks:**


**Related GitHub Issues:**