Add user friendly error for `autograph` bug with `qml.while_loop(lambda ...)` (#7178)
**Context:**

Lambda functions as loop conditions doesn't work with `autograph`
enabled. This is a very tricky bug to solve and so for the moment we
will just catch the ambiguous error it spits out and reform it into
something useful for the users.

Although, technically the user shouldn't even be using `autograph` with
`qml.while_loop`. 😄

**Description of the Change:**

Add a user friendly error message that tells the user how to fix it.

**Benefits:**

```python
qml.capture.enable()

def circuit():

    @qml.while_loop(lambda i: i < 3)
    def loop(i):
        qml.H(wires=i)
        return i + 1

    return loop(0)

>>> qml.capture.run_autograph(circuit)
```
```
AutoGraphError: AutoGraph currently does not support lambda functions as a loop condition for `qml.while_loop`. Please define the condition using a named function rather than a lambda function.
```

**Possible Drawbacks:** Bug isn't actually fixed. 

[sc-88411]