Removing import of  `DefaultTensor` in `init` file (#7193)
**Context:** To use the `default.tensor` device in PennyLane, we must
import the `quimb` Python package.

The strategy adopted for this was the same previously used for the
`default.clifford` device, which depends on the `stim` Python package.
That is, when the user imports `pennylane`, the package is imported if
installed on the system, otherwise it is not imported. An error is
raised if the user explicitly tries to instantiate the device without
the package installed.

However, `quimb` frequently raises a warning message when it is imported
if `kahypar` is not installed. This seems to be a long-standing issue
with the `contengra` package, and causes the following unpleasant
situation: if a user has `quimb` installed on the system (along with
some other packages) and imports `pennylane`, the following warning
message is raised:

```
UserWarning: Couldn't import `kahypar` - skipping from default hyper optimizer and using basic `labels` method instead.
```

**Description of the Change:** We no longer import the `DefaultTensor`
class in `pennylane/devices/__init__.py`.

**Benefits:** Now, the warning message is only raised if the user
explicitly instantiates the device, which makes more sense and is less
ambiguous:

```
qml.device("default.tensor", wires=2)

UserWarning: Couldn't import `kahypar` - skipping from default hyper optimizer and using basic `labels` method instead.
  warnings.warn(
```

Furthermore, this also speeds up the pennylane import process.

**Benefits:** Warnings are no longer raised when importing pennylane if
`cotengra` is installed along with other packages.

**Possible Drawbacks:** The documentation of the `DefaultTensor` device,
as listed in the 'next generation devices', is slightly different:


![DT_attempt](https://github.com/user-attachments/assets/a26142c0-e46d-4deb-9d89-9d15f0173636)

**Related GitHub Issues:** #6893 

[sc-83272]