Don't defer mid-circuit measurements when mcm_method is "device" (#7243)
**Context:**
Some mid-circuit measurement handling is device-specific, like
tree-traversal. When a method is device-specific, we want to skip
transforming the tape in the standard preprocess pipeline, and instead
rely on handling that is implemented in the `device` class.

This is currently the case for tree-traversal, but the more generic
specification `"device"` defaults to deferred measurements.

**Description of the Change:**

We handle `mcm_method="device"` in the `mid_circuit_measurements`
transform the same way we handle `mcm_method="tree-traversal"` - with a
null transform and `null_postprocessing`.

**Benefits:**
Generally, we can use `"device"` the way it seems to have been intended
as a method for handling mid-circuit measurements.

More specifically, we can use `"device"` as an MCM method on
`null.qubit` if we don't want to include transforming and
post-processing MCMs in our mock execution, but just want to treat MCMs
like any other operation on the tape with `null.qubit` and skip it. This
will be helpful for benchmarking the device processing and
post-processing for large-scale FTQC workloads on `null.qubit`.

[sc-89247]