add `exceptions.py` top-level file for custom PL errors/warnings (#7205)
**NOTE TO REVIEWERS**: My auto format on save impacted a few files -
mainly in the `tests` folder. Most likely these files were never
formatted with our formatting guidelines?

--- 

**Context:**

We have a few errors / warnings at top level for some reason. This is
bad because it is requiring `import pennylane as qml` in our code base.
This creates a lot of opportunities for circular dependencies.

**Description of the Change:**

The Plan: deprecate access from top-level. 🥴 

In this PR,

- Create a new `exceptions.py` file at top-level
- Copy over the top-level errors / warnings there
- Import at top-level from `exceptions` to retain feature parity

This leaves the errors accessible at top-level so we don't break any CI.
After merging this PR, follow-up PRs in our eco-system need to be
created to point to the new file,

- **Catalyst**: https://github.com/PennyLaneAI/catalyst/pull/1612
- **Lightning**:
https://github.com/PennyLaneAI/pennylane-lightning/pull/1122

**Plug-ins**

- **Qiskit**: https://github.com/PennyLaneAI/pennylane-qiskit/pull/626
- **Qulacs**: https://github.com/PennyLaneAI/pennylane-qulacs/pull/90
- **IonQ**: No changes required.
- **AQT**: https://github.com/PennyLaneAI/pennylane-aqt/pull/86
- **QRack**:
https://github.com/unitaryfoundation/pennylane-qrack/pull/31
- **Cirq**: https://github.com/PennyLaneAI/pennylane-cirq/pull/214

Once these PRs are merged, a follow-up PR in PennyLane is ready that
will remove the exceptions from top level,
- PennyLane: https://github.com/PennyLaneAI/pennylane/pull/7207

Done! 🥳 🎉 

**Benefits:** No more top level imports for these errors and warnings!

**Possible Drawbacks:** None identified.

[sc-88318] [sc-88950] [sc-89160]