Add informative error to capture + jit +  condition on abstract value + default qubit execution (#7314)
**Context:**

While many circuits can be jit-executed on default qubit with program
capture, we cannot yet handle executing abstract conditions. This adds a
more informative error to that effect.

If we decide that a full program capture pipeline with worth fully
developing and maintaining, we can come back to this, but for now, we
can just error out during device execution.

**Description of the Change:**

**Benefits:**

**Possible Drawbacks:**

**Related GitHub Issues:**