Update unit tests workflows to keep `dev` suffix for pennylane-lightning in frozen deps files (#7315)
Test workflows for updating stable deps were installing dev versions of
lightning, but the output file with the frozen deps showed that a stable
version of lightning was installed. This was because the step in the
workflow that created the frozen deps file was manually removing the
`.devX` suffix from the version of pennylane-lightning. This PR removes
the suffix removal from that step.