Remove keras layer (#7320)
**Context:**
We are to finish the deprecation cycle for `KerasLayer`. Note that this
does not mean that we are fully removing Keras from our path. There is
[another PR ](https://github.com/PennyLaneAI/pennylane/pull/7272)ongoing
to introduce a compatible version `KerasLayer` for Keras3; also, several
other modules depend on keras but not involving `KerasLayer`.

**Description of the Change:**
Deletion of source code file along with test file.
Removal of necessary special treatment in the middle of pipeline.

**Benefits:**

**Possible Drawbacks:**

**Related GitHub Issues:**
[sc-89146]

---------

Co-authored-by: Christina Lee <christina@xanadu.ai>
Co-authored-by: Pietropaolo Frisoni <pietropaolo.frisoni@xanadu.ai>
Co-authored-by: Isaac De Vlugt <34751083+isaacdevlugt@users.noreply.github.com>