Allow operators to have zero wires (#7327)
**Context:**

Maintaining a hardcoded list of concrete operators in an abstract base
class is a really smelly idea that makes it difficult to extend
`Operator` without touching it's source code. An interface should not
care about it's concrete implementations.

This is also blocking my catalyst PR:
https://github.com/PennyLaneAI/catalyst/pull/1676 since I need to make
`HybridOp` exist on zero wires. The `Operator` class should not care
about `HybridOp` or know of it's existance.

Moral of the story, the requirement that operators have at least one
wire is unnecessarily restricting.

**Description of the Change:**

Eliminates the check that operators need to exist on at least one wire.

**Benefits:**

I can update `HybridOp` to work on zero wires, so I can update
`Operator.num_wires` to default to `None`.

**Possible Drawbacks:**

**Related GitHub Issues:**

[sc-89159]

---------

Co-authored-by: Pietropaolo Frisoni <pietropaolo.frisoni@xanadu.ai>