The `cancel_inverses` provides a deterministic output with program capture (#7328)
Right now, the cancel_inverses transform with capture enabled uses a set
in its implementation to reshuffle operators temporarily stored by the
CancelInversesInterpreter.

Using a set causes the order to no longer be guaranteed, so the new
order of the
operations might differ from the original order, and in principle, it
can be different for different executions.

Although this only impacts operators without any shared wires (so
correctness will not be impacted) it would be good to have a transform
that provides a deterministic result. That is, we should know that the
output always provides the same result for the same input.

[sc-85610]

---------

Co-authored-by: Isaac De Vlugt <34751083+isaacdevlugt@users.noreply.github.com>