Capture parametric MCMs in plxpr (#7219)
Current implementation summary:
- We leave `ftqc.measure_z` completely equivalent to `qml.measure`, so
it captures the standard `measure` primitive for PennyLane

```
qml.capture.enable()

@qml.qnode(dev)
def circ1():
    qml.measure(0)
    return qml.expval(Z(0))
    
@qml.qnode(dev)
def circ2():
    measure_z(0)
    return qml.expval(Z(0))
```

```
>>> jaxpr = jax.make_jaxpr(circ1)()
>>> jaxpr.eqns[0].params["qfunc_jaxpr"].eqns
[_:i64[] = measure[postselect=None reset=False] 0,
 a:AbstractOperator() = PauliZ[n_wires=1] 0,
 a:AbstractMeasurement(n_wires=None) = expval_obs b]
```

```
>>> jaxpr = jax.make_jaxpr(circ2)()
>>> jaxpr.eqns[0].params["qfunc_jaxpr"].eqns
[_:i64[] = measure[postselect=None reset=False] 0,
 a:AbstractOperator() = PauliZ[n_wires=1] 0,
 a:AbstractMeasurement(n_wires=None) = expval_obs b]
```

- We capture `measure_x`, `measure_y` and `measure_arbitrary_basis` all
with the same parametrized measurement primitive, `p_measure`. This is
under the assumption that the distinction is primarily a user-facing one
for reading code and drawing circuits, as well as more specific
diagonalization, and none of those things are relevant at the point
where we are capturing for catalyst, so they may as well all just be one
thing that is handled in a uniform manner. So when captured, these
circuits are identical:

```
@qml.qnode(dev)
def circ1():
    measure_x(0)
    return qml.expval(Z(0))

@qml.qnode(dev)
def circ2():
    measure_arbitrary_basis(0, angle=0, plane="XY")
    return qml.expval(Z(0))
```

```
>>> jaxpr = jax.make_jaxpr(circ1)()
>>> jaxpr.eqns[0].params["qfunc_jaxpr"].eqns
[_:i64[] = p_measure[angle=0 plane=XY postselect=None reset=False] 0,
 a:AbstractOperator() = PauliZ[n_wires=1] 0,
 a:AbstractMeasurement(n_wires=None) = expval_obs b]
```

```
>>> jaxpr = jax.make_jaxpr(circ2)()
>>> jaxpr.eqns[0].params["qfunc_jaxpr"].eqns
[_:i64[] = p_measure[angle=0 plane=XY postselect=None reset=False] 0,
 a:AbstractOperator() = PauliZ[n_wires=1] 0,
 a:AbstractMeasurement(n_wires=None) = expval_obs b]
```

- We capture `cond_measure` by putting it inside a `cond` primitive
using `qml.cond`. This wouldn't work without program capture on, but my
understanding is that this will be more flexible with plxpr. We can make
a new primitive if that's not the case. The essential thing here is that
we need to be able to use the value returned by whichever function is
called going forward, so `cond` needs to have the option to return
something, i.e. in the following we need `a:i64[]` to be the value
returned by whichever of the measurement functions we end up executing
inside cond:

```
@qml.qnode(dev)
def circ():
    m1 = measure_x(0)
    m2 = cond_measure(m1, measure_x, measure_y)(2)
    cond_measure(m2, measure_x, measure_y)(3)
    return qml.expval(Z(0))
```

```
>>> jaxpr = jax.make_jaxpr(circ)()
>>> jaxpr.eqns[0].params["qfunc_jaxpr"].eqns[1]
a:i64[] = cond[
  args_slice=slice(2, None, None)
  consts_slices=[slice(2, 2, None), slice(2, 2, None)]
  jaxpr_branches=[{ lambda ; a:i64[]. let
    b:i64[] = p_measure[angle=0 plane=XY postselect=None reset=False] a
  in (b,) }, { lambda ; a:i64[]. let
    b:i64[] = p_measure[
      angle=1.5707963267948966
      plane=XY
      postselect=None
      reset=False
    ] a
  in (b,) }]
] b True 2
```

- We make a new function, `make_graph_state`, that either creates a
GraphStatePrep operation (if `capture` is off), or decomposes it and
captures the operations generated in the decomposition (if `capture` is
on). This circumvents JAX's inability to handle a networkx graph. Maybe
we want to maintain some more efficient way of storing this in a single
primitive in the future, but for now this seems simple and acceptable.

```
import networkx as nx
from pennylane.ftqc import make_graph_state

@qml.qnode(dev)
def circ():
    make_graph_state(nx.grid_graph((4,)), wires=[1, 2, 3, 4])
    return qml.expval(X(4))
``` 

```
>>> jax.make_jaxpr(circ.func)()
{ lambda ; . let
    _:AbstractOperator() = Hadamard[n_wires=1] 1
    _:AbstractOperator() = Hadamard[n_wires=1] 2
    _:AbstractOperator() = Hadamard[n_wires=1] 3
    _:AbstractOperator() = Hadamard[n_wires=1] 4
    _:AbstractOperator() = CZ[n_wires=2] 1 2
    _:AbstractOperator() = CZ[n_wires=2] 2 3
    _:AbstractOperator() = CZ[n_wires=2] 3 4
    a:AbstractOperator() = PauliX[n_wires=1] 4
    b:AbstractMeasurement(n_wires=None) = expval_obs a
  in (b,) }
```

```
>>> qml.capture.disable()
>>> qml.workflow.construct_tape(circ)().operations
[GraphStatePrep(Hadamard, CZ)]
```
---------------------------------

**Context:**

**Description of the Change:**

**Benefits:**

**Possible Drawbacks:**

**Related GitHub Issues:**


[sc-88520]

---------

Co-authored-by: Joey Carter <joseph.carter@xanadu.ai>